refactor: 사용자 회원가입 책임 다시 이동
