+++ src/main/java/roomescape/member/domain/Role.java
@@ -1,5 +1,9 @@
 package roomescape.member.domain;
 
 public enum Role {
-    ADMIN, GUEST
+    ADMIN, GUEST;
+
+    public boolean isAdmin() {
+        return this == ADMIN;
+    }
 }
