+++ src/main/java/roomescape/member/service/MemberService.java
@@ -26,16 +26,4 @@ public class MemberService {
     public Member findById(long id) {
         return memberRepository.getById(id);
     }
-
-    public MemberResponse create(SignUpRequest signUpRequest) {
-        validateEmail(signUpRequest.email());
-        Member member = memberRepository.save(new Member(signUpRequest.name(), signUpRequest.email(), signUpRequest.password(), Role.GUEST));
-        return new MemberResponse(member);
-    }
-
-    private void validateEmail(String email) {
-        if (memberRepository.existsByEmail(email)) {
-            throw new InvalidMemberException("이미 가입된 이메일입니다.");
-        }
-    }
 }
