+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -30,4 +30,7 @@ public class AdminReservationController {
         reservationService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
+
+  /*  @GetMapping("/search")*/
+
 }
